# Backend/MyBackend/Models/LeaveSessionBody.cs
namespace MyBackend.Models
{
    public class LeaveSessionBody
    {
        public string UserName { get; set; } = string.Empty;
        public string SessionTitle { get; set; } = string.Empty;
    }
}