# Backend/MyBackend/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "MyBackend, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/players/playersService": {
      "get": {
        "tags": [
          "Players"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "title",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/PlayerForService"
              }
            }
          }
        }
      }
    },
    "/api/players/leaveSession": {
      "patch": {
        "tags": [
          "Players"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/LeaveSessionBody"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/sessions/{sessionId}": {
      "get": {
        "tags": [
          "Sessions"
        ],
        "operationId": "GetPlayer",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/SessionModel"
            }
          }
        }
      }
    },
...
    "/api/sessions/sessionPicture": {
      "get": {
        "tags": [
          "Sessions"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "title",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/sessions/addSession": {
      "post": {
        "tags": [
          "Sessions"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/SessionPost"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ParamsAddSessionResponse"
            }
          }
        }
      }
    },
    "/api/sessions/addPlayerToSession": {
      "patch": {
        "tags": [
          "Sessions"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AddPlayerToSessionBody"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/sessions/deleteSession": {
      "delete": {
        "tags": [
          "Sessions"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "sessionTitle",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
    "AddPlayerToSessionBody": {
      "type": "object",
      "properties": {
        "sessionTitle": {
          "type": "string"
        },
        "playerName": {
          "type": "string"
        }
      }
    },
...
    "GameSystemModel": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        }
      }
    },
...
    "GmModel": {
      "required": [
        "name"
      ],
      "type": "object",
      "properties": {
        "name": {
          "maxLength": 50,
          "minLength": 1,
          "type": "string"
        },
        "sessions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SessionModel"
          }
        }
      }
    },
    "LeaveSessionBody": {
      "type": "object",
      "properties": {
        "userName": {
          "type": "string"
        },
        "sessionTitle": {
          "type": "string"
        }
      }
    },
...
    "ParamsAddSessionResponse": {
      "type": "object",
      "properties": {
        "isTitleUnique": {
          "type": "boolean"
        },
        "isSystemNew": {
          "type": "boolean"
        }
      }
    },
...
    "PlayerForService": {
      "required": [
        "name"
      ],
      "type": "object",
      "properties": {
        "name": {
          "maxLength": 50,
          "minLength": 1,
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "profilePicture": {
          "type": "string"
        },
        "discord": {
          "type": "string"
        }
      }
    },
...
    "PlayerModel": {
      "required": [
        "name"
      ],
      "type": "object",
      "properties": {
        "name": {
          "maxLength": 50,
          "minLength": 1,
          "type": "string"
        },
        "sessions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SessionModel"
          }
        }
      }
    },
...
    "SessionModel": {
      "required": [
        "description",
        "title"
      ],
      "type": "object",
      "properties": {
        "system": {
          "$ref": "#/definitions/GameSystemModel"
        },
        "title": {
          "maxLength": 50,
          "minLength": 1,
          "type": "string"
        },
        "tags": {
          "type": "string"
        },
        "description": {
          "maxLength": 500,
          "minLength": 1,
          "type": "string"
        },
        "maxNumberOfPlayers": {
          "format": "int32",
          "type": "integer"
        },
        "gm": {
          "$ref": "#/definitions/GmModel"
        },
        "players": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/PlayerModel"
          }
        }
      }
    },
    "SessionPost": {
      "type": "object",
      "properties": {
        "system": {
          "$ref": "#/definitions/GameSystemModel"
        },
        "title": {
          "type": "string"
        },
        "tags": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "picture": {
          "type": "string"
        },
        "maxNumberOfPlayers": {
          "format": "int32",
          "type": "integer"
        },
        "gmUserName": {
          "type": "string"
        }
      }
    },
...
  }
}